+++ src/ru/itmo/potatocoder228/androidlab2024/database/interfaces/HouseCollection.kt
@@ -2,8 +2,8 @@ package database.interfaces
 import model.House
 public interface HouseCollection{
     fun save(house: House)
-    fun findByUserId(id: Int) : House
-    fun findById(id: Int) : House
+    fun findByUserId(id: Long) : House
+    fun findById(id: Long) : House
     fun update(house: House)
 
 }
