+++ src/ru/itmo/potatocoder228/androidlab2024/model/User.kt
@@ -2,5 +2,5 @@ package model
 public data class User(
     var login: String,
     var password: String,
-    var id: Int = 0,
+    var id: Long = 0,
 )
